Removed _hexbin directory
